Fixed the Change

+++ EpicMarket.Data/Models/Outlet.cs
@@ -23,8 +23,8 @@ namespace EpicMarket.Data.Models
         public string Description { get; set; }
         public long ContactNumber { get; set; }
         public string ContactEmail { get; set; }
-        public int? Rating { get; set; }
-        public int? ReviewCount { get; set; }
+        public double? Rating { get; set; }
+        public double? ReviewCount { get; set; }
         public bool IsOpen { get; set; }
         public double? Weight { get; set; }
